Fix division in VProgressBar constructor to ensure correct tick calculation

+++ src/v_progress_bar.cpp
@@ -21,7 +21,7 @@ VProgressBar::VProgressBar(const VProgressBarOptions& options)
     if (chars.size() < 2) {
         throw std::invalid_argument("char_frames must have exactly 2 elements (for empty and filled states), got " + std::to_string(chars.size()));
     }
-    tick = 100.0 / (chars.size() - 1);
+    tick = 100.0 / (static_cast<double>(chars.size()) - 1);
     showCursor(false);
     redraw();
 }
